user/comそれぞれのメソッドを役割別に振り分け
